firmware: declare platform packages to use for esp32 libraries
Lock the platform to the pioarduino espressif32 fork at version
 3.1.0.
Building with the official espressif32 plattformio provides we are
looked to the arduino core v2.x.x even when instructed to use newer
libraries.
Use the fork for now to build the project with the arduino core
v3.x.x.