firmware: intervalometer: refactor and abstract functions into own classes
Remove the obsolet night timelapse functionality as it is already contained
in the day timelapse. Unify it into a single timelapse mode.
Now this introduces a new mode where we pan the mount during timelapse
shooting. Currently we only pan while the delay.
This mode is useful for shooting panoramas or startrails.