firmware: axis: fix wrong direction of goto
This is an attempt to fix the goto functionality to move into the
right direction. It uncovered alot of work below to clarify used
variables in the context.

What was actually necessary to make goto pan into the right direction
is overwriting the absolute direction, setting the drivers
direction directly.
Furthermore goto didn't stop when the target was reached as the counting
did not behave as expected with a bunch of microstepping conversion.